rename metadata.ts to property.ts
