+++ packages/forms/experimental/public_api.ts
@@ -14,7 +14,7 @@
 export * from './src/api/async';
 export * from './src/api/control';
 export * from './src/api/logic';
-export * from './src/api/metadata';
+export * from './src/api/property';
 export * from './src/api/structure';
 export * from './src/api/types';
 export * from './src/api/validators';
