+++ packages/forms/experimental/src/api/validators/pattern.ts
@@ -7,7 +7,7 @@
  */
 import {computed, Signal} from '@angular/core';
 import {aggregateProperty, property, validate} from '../logic';
-import {PATTERN, Property} from '../metadata';
+import {PATTERN, Property} from '../property';
 import {FieldPath, LogicFn, PathKind} from '../types';
 import {ValidationError} from '../validation_errors';
 import {BaseValidatorConfig} from './util';
