+++ packages/forms/experimental/src/field/property.ts
@@ -7,7 +7,7 @@
  */
 
 import {computed, runInInjectionContext, Signal, untracked} from '@angular/core';
-import {AggregateProperty, Property} from '../api/metadata';
+import {AggregateProperty, Property} from '../api/property';
 import {FieldNode} from './node';
 import {cast} from './util';
